always compare against stringified object when pattern is RegExp
When the pattern is a regular expression and the object has not
already been detected as a regular expression too, one can be almost
100 % certain that the user meant to compare the stringified version
of the left-hand side (e.g. for errors, arrays, etc.) against
the regular expression.

Previously, this would almost always pass silently in these simple
cases.

This is likely a semver-major change.